pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sqlmentor"
version = "0.1.0"
description = "SqlMentor — CLI para coleta de contexto Oracle para tuning de SQL assistido por IA"
requires-python = ">=3.9"
dependencies = [
    "typer[all]>=0.9.0",
    "sqlglot>=20.0.0",
    "oracledb>=2.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "mcp[cli]>=1.0.0",
]

[project.scripts]
sqlmentor = "sqlmentor.cli:app"
sqlmentor-mcp = "sqlmentor.mcp_server:main"

[project.optional-dependencies]
dev = [
    "hypothesis>=6.0",
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "taskipy>=1.12.0",
    "ruff>=0.4.0",
]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "S", "PT", "SIM", "RUF"]
ignore = [
    "E501",   # line too long — já controlado por line-length
    "B008",   # function call in default argument (Typer usa isso)
    "B904",   # raise dentro de except sem from — padrão legítimo em CLI Typer
    "S101",   # assert em testes — pytest usa assert
    "S603",   # subprocess sem shell=True — não usamos subprocess
    "S607",   # partial executable path — não aplicável
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["F811", "S101", "S105", "PT011", "RUF002", "RUF003", "RUF005"]
"src/sqlmentor/report.py" = ["RUF001", "RUF002", "RUF003"]  # unicode intencional em relatórios (×, –, ℹ️)

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: requer conexão Oracle real",
]

[tool.coverage.run]
source = ["sqlmentor"]

[tool.coverage.report]
show_missing = true

[tool.taskipy.tasks]
test     = "pytest tests/ -v"
test-cov = "pytest tests/ -v --cov=sqlmentor --cov-report=term-missing"
test-run = "pytest tests/ -v -p no:randomly"
test-q   = "pytest tests/ -q"
lint     = "ruff check src/ tests/"
lint-fix = "ruff check src/ tests/ --fix"
format   = "ruff format src/ tests/"

[tool.setuptools.packages.find]
where = ["src"]
